test info

+++ alloc_fast.go
@@ -8,7 +8,7 @@ import (
 )
 
 func newNodeFast(item Item, level int) (node *Node) {
-	node = (*Node)(C.malloc(int(unsafe.Sizeof(Node{}) + uintptr(level)*unsafe.Sizeof((*Node)(nil)))))
+	node = (*Node)(C.malloc(unsafe.Sizeof(Node{}) + uintptr(level)*unsafe.Sizeof((*Node)(nil))))
 	h := (*reflect.SliceHeader)(unsafe.Pointer(&node.forward))
 	h.Data = uintptr(unsafe.Pointer(node)) + unsafe.Sizeof(Node{})
 	h.Cap = level
